Added foodorders to keep track which ant is going for which food and to prevent two go for the same food. It is collected before the food moves are issued, so right in the beginning of every turn.
Changed getClosestFood for one ant so it skips the food, that's taken care of, if the other ant is closer to it

Removed check for maximum distance to food. doesnt make sense, if food is visible, then because of a close ant. That close ant should take care of and this should work by the check in isAnotherAntCloserToThisFood

Refactored the method to find an ant by Location, so its in just one place.
introduced Targets for Orders.